+++ module/actors/sheets/director.mjs
@@ -246,4 +246,6 @@ export class OutgunnedDirectorSheet extends ActorSheet {
       await this.actor.update({ [`system.${collectionName}`]: collection })
     }  
   }
+
+
 }